requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
dnspython==2.6.1
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pycparser==2.22
pymongo==4.7.3
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.2
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
